[Style] CoupleException 파일 통합, CoupleIllegalArgumentException 추가
