--- src/main/kotlin/com/whatever/domain/couple/exception/CoupleAccessDeniedException.kt
@@ -1,6 +1,0 @@
-package com.whatever.domain.couple.exception
-
-class CoupleAccessDeniedException(
-    errorCode: CoupleExceptionCode,
-    detailMessage: String? = null
-) : CoupleException(errorCode, detailMessage)
