--- src/main/kotlin/com/whatever/domain/couple/exception/CoupleIllegalStateException.kt
@@ -1,6 +1,0 @@
-package com.whatever.domain.couple.exception
-
-class CoupleIllegalStateException(
-    errorCode: CoupleExceptionCode,
-    detailMessage: String? = null
-) : CoupleException(errorCode, detailMessage)
